move unit and category to new files
